End Frozen Ramparts upon level change (Lightli)
The implementation for level cleanup when the Ramparts status expires
wrongly assumed that the player is still on the current level. They
could, of course, leave the level and have the status expire on the new
level, returning to the old level later. This meant that walls could
become permanently be flagged as icicle covered, even though they would
only operate as such if the player was in LOS and had the spell active.

This commit has the Ramparts status expire and trigger the usual feature
property cleanup whenever a level change happens.