Rename shield -> hoplon
As explained in ca9fc5ac96d, we'd like to be able to refer to an
item without confusing it with a category. A hoplon was the shield
used by a hoplite, a 'medium' size similar to the ones in-game.

Alternatives considered were:
- targe (too close to a buckler in both size and meaning)
- aspis (the 'proper' name for a hoplon, but come on)
- kite (silly & specific)
- heater (see kite)
- dhal (again, too close to a buckler)
- umbumbuluzo (cool but way too long)
- ihawu (cool but it's three syllables)

I think the closest alternative contender was 'targe', fwiw.

The shield of resistance and of the gong were intentionally not
renamed, which I don't believe should cause confusion, but we can
rename them later if it does.