New single-use item: Tremorstone
Tremorstones are single-use evocables that cause a radius 2 explosion at
range 3. Their explosion is skewed by 1 tile from the target, so they always
have at least a 1/3 chance of hitting the player. However, since they do
6d12 BEAM_FRAG damage, players with 20-25 AC will usually take little to
no damage. (BEAM_FRAG checks 3x AC.) As such, this is intended to be a fun
toy for well-armoured characters and a desperation option for others.

A design space that I've wanted to explore for many years is the
'3x AC' damage type (ala LRD and the late, lamented Orb of Fragmentation).
I initially tried to make it into a spell, but it's very much the opposite
of the modern Earth school philosophy and in general makes more sense as a
tool for armoured, less-spell-focused characters.

This will always warn about self-targeting, which is a bit clunky. Since
it's not especially common and it's nice to warn players who might not have
read the description before they blow themselves up, though, that seems OK.