Remove fans of gales
Miscellaneous evocable items are currently a very overloaded item type, with
many items that are worth every character carrying once found, as well as a
lot of duplicated effects. This results in lots of inventory slots being
dedicated to evokers, a worse interface when picking items, and a lot of
uninteresting choices between similar items. Narrowing down the list of evokers
substantially will allow them to have a smaller number of distinct effects,
without taking up so much inventory load.

The fan of gales AoE knockback is a very niche escape effect, and isn't really
interesting enough to dedicate an item to - it falls into the category of
evokers that end up always being carried around even though in practice it's
only quite rarely useful.