Clean up icy walls upon level load
Frozen Ramparts creates "icy walls" with the FPROP_ICY property. It
expires the props when the Ramparts duration expires, but early versions
of the spell didn't do this properly. In this commit we unset the prop
for all squares with the property that don't that aren't properly in LOS
of the last position of any active cast of Frozen Ramparts. This will
fix walls changed by buggy versions of the spell upon level load. This
cleanup is tag-majored, since it shouldn't be necessary for versions
where the spell logic is fixed.

Additionally, the icy wall presence level state is now updated properly
upon level load when any valid icy walls are found for a current cast of
Frozen Ramparts. Previously this wasn't done, so casts active at time of
save didn't have working icy walls when the save was reloaded.