Mark potions of agility as removed (Yermak)
It's necessary to add these potions to the removed_items set so that we
don't get a "potion of bugginess" entry in the unknown item listing of
the identification menu. This entry might also be necessary to disable
things like generation, but I haven't checked this.

A subsequent commit should TAG_MAJOR various portions of agility potion
code, the related agility duration code, and remove the effect from the
list of Xom effects.