Assimilate handling of dummy agility to other removed potion types
This is unfortunately ugly, but it does not crash. Basically, treat this
very temporary potion like any other potion that's been removed -- even
though it's last in the enum, keep it around until a major version tag,
and do save compat unconditionally.

This is a semi-revert of 08029b5080ff (sorry).