Refactor some regen code
Factor out the logic for activating a prop if some condition is met,
which was repeated several times for activating regeneration, mana
regeneration, and acrobat amulets when hp or mp maxes out.

This function isn't used anywhere else, and that seems unlikely to
change soon, so it's now a static function in player-reacts.cc.