Make TLA and regen artefacts activate after maxing hp
Currently it makes sense to swap to troll leather armour or armour with
the regen+ artefact property when resting, provided one is going to rest
long enough and can do so in a safe enough position. This is because the
regen allows one to rest faster, helping the player deal with the clocks
in the game (piety decay, nutrition, portal timers, score). Although in
practice this seems to mostly be the province of speed-runners, it's
often also a tedious-but-optimal tactic for normal play.

This commit makes that tactic unavailable by extending the current
disincentive for swapping amulets of regeneration to regen armour.
Regeneration items now only affect regen rate after the player has
healed to max hp with the item on. Thus benefiting from regen requires a
long-term equipment choice.

Deactivating a regen item on unequip is moved to an unequip_effect
instead of handling it in player_reacts, on the principle that status
changes that happen due to unequiping an item should probably happen
there.

This commit makes it so melding a regen item stops the regen bonus, but
does not deactivate the item, i.e., you don't have to rest up to full to
get the effect back after untransforming.

[ Committer's note: added artefact names to messaging, refactored a
  switch statement to use comma_separated_line from stringutil instead ]