Rework the potion of brilliance
At present, the potion of brilliance is an odd bird. It decreases
spell failure cost dramatically and increases spell power slightly,
essentially acting as a 'potion of cast high fail chance spells'.
This encourages players to memorize spells that they can't normally
cast just for the few times during a game they'll use !brill. The
UI for this isn't great (how do you tell what your fail chance will
be after you drink it?), and it's relevant to almost all characters,
meaning that everyone's inventories and tactical options are
cluttered with it.

This changes !brilliance to instead do two specific things: act as
a universal spell enhancer (like archmagi) and reduce spell hunger
costs by half. The former is a buff for 'casters', but probably
not something most 'melee-focused' characters care about (and thus
need to have in their inventory). The latter is intended to act as
moral encouragement to blast away with your most powerful spells
while the potion is running.

The new brilliance looks more powerful than the old one, so its
duration is roughly halved.