Disallow -tele on artefact weapons and jewellery
-tele is usually not a meaningful bad artefact property on weapons and
jewellery, since swapping away from those items is so quick.
Furthermore, putting -tele on quick-swapping items allows the player to
explore with a -tele item equipped and swap to their "real" equipment
for fights, thereby avoiding teleport traps and teleportitis at marginal
cost. This is frustratingly uninteresting gameplay. And it allows the
player to neutralize the gameplay benefits of teleport traps with little
risk and no real investment or decision-making regarding their
character's loadout.

A simple solution is to prevent -tele from spawning on any weapons or
jewellery. We could also try to make sure that on jewellery and weapons
-tele always pairs with other properties that make swapping less
trivial (*contam, fragile, harm, etc.), but I am skeptical that the
complexity is worth the trouble.

I also considered changing the behaviour of -tele to stop translocations
only from an internal source. Then wearing -tele would not affect
teleporting unless the player could already choose whether or not to
teleport, so there is no reason to swap to -tele to explore. However
this is additional "player-facing" complexity, in that there is another
somewhat fiddly rule for the player to learn; even very careful players
don't need to know about rules governing placement of artefact
properties. Furthermore it does not address the first issue mentioned,
that weapon and jewellery -tele is rarely an interesting drawback.
Finally this approach takes away an arguably interesting strategic
choice that players are making, to wear -tele on armour to prevent
teleport traps in particular places (esp. Zot).

[ Committer's note: simplified and added a comment. Closes #1292 ]