Let scroll of immolation affect summons
Since this scroll was added in 388d670c250b25dcf88b5f54042bb083f48a9a51,
the effect has been limited to non-summoned creatures. It's hard to be
exactly sure why.

This limitation is a little non-intuitive, and for a limited resource
like scrolls, probably unnecessary. The scrolls are situational enough
as-is.