Remove the str bonus from !might
!might is a potion that makes you much better at melee combat. It
also, incidentally, gives you 5 str, which adds a little more melee
damage, improves your accuracy and spellcasting chances in armour,
and so on.

Simplify and focus on the core idea of the potion by removing the
bonus strength. The bonus damage remains very beefy and noticable,
and there are plenty of other ways to play with HUGE STRENGTH.