Add room for tactics to vaults in twisted.des
I also simplified a few of the vault definitions by removing meaningless
statements, etc.

Vault list:

- shiori_arrival_whats_in_store_for_you

This vault has a single path and half of the time a single exit, which
makes it possible for the player to get trapped by a bad monster spawn
outside the vault. I made it always have 2 exits.

- dpeg_arrival_solitude

This vault has one of three single-exit paths. I added a second exit to
each. I also simplified a shuffle/subst sequence that had meaningless
features.

- dpeg_arrival_connections

This vault could potentially spawn you in a corner with no paths behind
you; depending on what the randomly generated terrain and monsters are,
that could be pretty bad. I added an escape hatch in whichever quadrant
the player arrives in.

- shiori_arrival_maze

It's possible for this vault to only give you exits from the starting
chamber to a single corner, and then only be able to leave that room to
the outside through a singe exit. I added a second exit to each corner
room. I also removed a meaningless SHUFFLE command.

- erik_arrival_cave_to_civilisation

The only guaranteed paths away from each starting location are
single-tile-wide single-path corridors for quite a ways. I added a
second internal path near each starting location to give some tactical
space.

- shiori_arrival_elements

Another single exit vault. There's not really a way to add a second exit
without screwing with the theme of this vault, so I added an escape
hatch to the starting room.

- nagdon_arrival_double_spiral

Another single-exit, single-path vault. I removed a couple of walls in
the center to provide some tactical depth while preserving the in/out
spiral symmetry. I also removed a KFEAT statement with no corresponding
glyphs.

- nagdon_arrival_zigzag_paths

There are two versions of this vault which were single-path,
single-exit. I added a second exit to each.

[ Committer's note: wrapped message to 80 cols. Closes #1299 ]