Change wand part1: Remove wand of enslavement/paralysis
There are too many wands in the game, so I'm going to organize it. The goal of cleanup is to unite the wands with similar roles into one. I will remove destructive effects from the wand of random effects, leaving only hex effects. Enslavement and paralysis will only be seen in wand of random effects in the future.

I will organize the spell list of wand of random effects in Part 2.