Hexify of wand of random effects
The immediate effect of damaging the enemy is provided by the other
wands. The old effect list was at one point the union of wand effects,
but that is long gone and far from true, which makes it a bit of a
spoiler trap.

The new effect list is entirely buffs and debuffs, and listed in the
wand description.

If the power level isn't right the effects can be weighted and tuned.

[ Committer's note: Closes #1310. Revised the effect list, re-worded the
  commit message. ]