Properly set the admin flag in the WebTiles client
Some testing code was leftover from debugging the client-side portion of
the new admin panel such that all users saw this panel regardless of
actual admin status. The user admin check was properly handled in the
WebTiles server, so no non-admins could successfully use the message
facility. This bug just affected the visibility of the panel in the
client lobby.