A player ghost level for Ziggurats
This commit adds a player ghost monster set for Zigs that places ghosts
of players who've died in a Zig. The rest of the monsters have suitable
undead theme:

* soul eaters for LOS-wide HP drain and spectral warriors to remove MR
  (moderate weight decreasing with depth).
* deep elf death mages and ghost crabs to heal their fellow undead
  (flat, low weight).
* shadow dragons for ranged and melee draining as well as trampling
  (flat, low weight).
* eidolons for af_slow and revenants for dispel undead (increasing up to
  a moderate weight).
* black suns for more dispel undead and slowing as well as antimagic
  (flat, low weight).
* curse skulls and curse toes for torment and bogging players down with
  summons (flat, low weight).

This gives a nice mix that's light on ghosts in early depths but ramps
up to 5 or more by halfway through the Zig. The player ghost set
replaces the "draining" monster set, since draining monsters are well
represented by this new one along with the Tartarus and Crypt sets. This
set remains at weight 2 to keep the frequency of "undead" floors from
being too high.

TODO: The Zig permastore bones file we ship has 45 ghosts, which is
already over the max of 20 we allow for adding new ghosts. Since a
server's permastore can thus only replace ghosts and won't grow past 45,
and since the non-permastore bones file is even smaller, player ghost
Zig levels are likely to see repeated ghosts. This is especially true of
mega Zigs, where monster counts of 650+ mean that a given ghost will be
repeated 12 times on-average. It might be good to have a higher limit
for the Zig bones files specifically, and to perhaps save XL27
characters that die in extended branches in the Zig file. This is
especially the case since we aren't currently using these extended
ghosts, although that's a separate issue to address.