Some tweaks to not_door_vault_mu
Make this whimsical version of door_vault place over the same range as
the actual door_vault. This way there aren't spoilers as to which
branches can give you an easy door vault. Use the shared setup function,
which also means we place loot. These vaults place late in the game and
this variant is weight 1, so some free loot in exchange for the bravery
of opening a door vault is fine. Finally, add a variant for Vaults
branch as an "empty" room.