Revamp door_vault
This is one of the great spoiler vaults that we allow to remain in the
game, but the vault could use some improvements. Expand its placement a
little to include D:13-14 instead of just D:$. Subsequent commits will
reduce the number of spoilers it does have by giving it and its
look-alike variants the same placement, for the most part. That way, no
matter where you open door vault, you may discover an 8 and 9 spawn zoo,
a hellephant zoo, or just a room full of lovely butterflies!

Tweak the monster sets to have uniformly difficult branch-specific
monsters together with end-level branch spawns at somewhat higher
weight. Before this change, the monster lists had individual weaker
monsters in addition to a high-weight normal branch spawn entry, which
is redundant. I've also made the nasty spawns more memorable, with
attention given to the top-tier branch-unique monsters.

Finally, this commit adds a common setup function for all door vault
variants to set loot and door markers and glyphs. Door vault was
notorious for not placing loot, making it purely an exercise in XP gain.
The addition of on-average 8 good items makes the decision to open a
door vault feel less silly.