Revamp not_door_vault_due
This "elephant" door vault variant placed normal and dire elephants, but
notably placed a single hellephant. Unfortunately, non-demonic elephants
are not challenging through the range of this vault's placement, and a
single hellephant does make up for this. So rather than leaving this as
an easier door vault variant, I've revamped the monster set to focus on
its hellish and demonic themes. I've also brought its level placement in
line with the other door vault variants to reduce spoilers.

The new monster set includes hell hogs and hell hounds for Dungeon, a
chaotic theme for Elf and Zot, and a mix of hell knights, hellions and
balrugs for other branches. At least one hellephant is always present,
and additional ones are added, scaling with depth of placement, up to 4
guaranteed hellephants in Geh and Dis.

As I've done for the other door vault variants, the common setup
function is used, which adds appropriate loot. There is also now a
version of this vault for the Vaults branch.