Add a player ghost door_vault variant
This door vault places multiple player ghosts, 7.5 on-average and up to
10. This will make an extra challenging yet rare door vault variant
that's sure to delight fans of gammafunk_ghost_cemetary. Accompanying
the ghost are 8 and 9 spawns, similarly to what is done for the original
door vault.

It uses the same common setup function as other door vaults, hence
places multiple loot items, but additionally places two more specialized
loot items that more aggressively upgrade in quality with depth, as it
done for other challenging ghost vaults. There is a ghost room variant
for Vaults to go along with variants for this branch added in prior
commits.

The weight of these variants is 2 both because they're extra challenging
and to keep placement less than the main door vaults.