Fix recasting after Divine Exegesis (AliasTheSpectator)
It's possible to cast a spell under Sif's Divine Exegesis and then use
the spellcasting interface to recast the spell using the '.' key even
when the spell isn't memorized. Spells cast this way don't get the
benefits of Divine Exegesis, so this bug only lets you save spell levels
by casting unmemorized spells you'd otherwise already be able to cast.

This commit fixes the spellcasting UI to properly unset the value of the
last cast spell if it isn't memorized before the spell prompt is made,
preventing the exploit.