Simplify and rebalance Frozen Ramparts
This spell nicely fills the gap left by the removal of Throw Frost, but
it suffers from being a mashup of two ideas: mine was a wall-oriented
positional spell and ebering's was a spell to slow monsters in some
interesting way. The former idea is relatively straightforward for a
damage spell, but slowing monsters has the potential pitfall of
introducing too much kiting. There's room for a spell that does both,
but this is harder to balance for the low-level spell we're going for.
Right now the slowing aspect of Ramparts isn't interesting, and the
spell is very strong while being in the strongest starting book in the
game.

This commit simplifies and rebalances the spell by removing the slow
movement monster debuff and making the damage pure BEAM_COLD. Hence
Ramparts is no longer partly resistable, but it will slow cold-blooded
things like hydras. I think a mass-freeze spell that uses walls is
interesting and reasonably balanced at level 3.

I've also tweaked the damage message to include damage punctuation and
to always use singular "wall" in the message. Many players aren't aware
that only one wall can ever effect a monster regardless of how many icy
walls are adjacent. Incidentally, this is also true of slimy walls
burning monsters, but the player *can* be affected by multiple such
walls. Monsters, they're dirty cheaters!