Rework trap placement in a Volcano map (SFR)
The map volcano_village placed an alarm trap in a mandatory passage near
the start of the map. Move this to an area before the guaranteed loot
room so that the player has more of a decision for triggering the traps.
Also reglyph the map and rework monster placement a little; it was
attempting to place a monster glyph that isn't defined, since the
"villager" set for the fiery_guardians() function only makes two monster
types.