Adjust spawn locations of D:1 monsters
Some vaults are small and place transparent rock, which means level
generation monsters can spawn in LOS of players. Although a very large
majority of these situations can be survived, it's not great to have such
a monster onslaught from turn 1. The function _zap_los_monsters() deals
with this by removing monsters that spawn with line of fire, but this
both reduces monster count and doesn't handle monsters that have
line-of-sight. Extending said function to line-of-sight would further
remove monsters, so a better approach is to prune D:1 monster spawn
locations without pruning the monsters themselves.

This commit introduces PROX_AWAY_FROM_DUNGEON_ENTRANCE and applies it to
level monster spawn locations on D:1. Any D:1 level monsters will not
choose a spawn location within LOS distance of the player, choosing
another location instead. This is just a strict distance check, not
considering actual line-of-site to the player, to avoid situations where
a monsters are very close to the starting position of the player, but
just around a corner or door in the starting vault. This change doesn't
handle monsters from any vaults that happen to generate too close. That
would very rarely be an issue, and we leave _zap_los_monsters() in place
to handle this reasonably. Objstat runs before and after this change
confirm that D:1 monster counts and XP are not affected by this change.

Now players who get the infamously small
scummos_arrival_simple_fountains, recreated here for your viewing
pleasure, can be glad they'll no longer be potentially surrounded by a
swarm of instantly awakened monsters:

.......
.T...T.
..mmm..
..m{+..
..mmm..
.T...T.
.......

If there are concerns about D:1 being easier, we might look into
providing some more interesting difficulty by adding some spicy rare
spawns to the D:1 monster table.