Reduce the range and damage of Frozen Ramparts
This spell's current radius of 3 can create a quite large number of
squares in typical use cases: 9 in corridors and potentially much more
than this when used at junctions of multiple paths. Reduce the radius of
effect to 2, which still gives plenty of affected squares.

Also reduce the portion of the damage affected by spellpower by 20%. At
a max spellpower of 50, it will slightly exceed the damage of Freeze.
These changes will bring the damage per cast a bit more in line for a
single school, level 3 spell.