Refactor and revise monster target acquisition
set_nearest_mosnter_foe is very old, it has its own duplication of a
distance_iterator. This commit refactors the function to use a
distance_iterator and adds a LOS restriction to cells considered.

As a consequence, ball lightning and foxfire no longer attempt to get at
monsters on the other side of clear rock.