Fix some issues in a vault
The vault gammafunk_dig_for_victory used light blue coloring for
transparent stone, but this is the same color used for grates. Remove
most of the wall recoloring, using it only inside the vault for
transparent stone that's surrounded by transparent rock. This recoloring
helps to draw attention to which tiles are diggable without looking odd
from the outside.

Supplement the wands charges in the supplied wand of digging so that it
will have enough to dig a reasonable number of times if the player is
going through the vault carefully. Also supplement the loot with more
good potions and scrolls, so that we're more likely to tempt people in.

Finally, remove no_item_gen, since it's not necessary, fix up the
no_tele_into KPROP, and realign the DES statement.