Adjust Vehumet gift range
Based on discussion in Mantis ticket 12221, this tweaks the range
of Vehumet spell gifts to try to avoid repeated offers at gifts
number 10, 11 and 12.

Extending the range from 6-7 to 5-7 is a small nerf, however
level 5 contains desirable spells (Fireball, LRD). To compensate,
range of gift 12 (the last one before the final) is extended
from 6-7 to 6-8.