Adjust Trog wrath
Remove the chance for rotting; while it was intended as "physical
damage" (judging from the comments) rotting is necromantic.

Replace the fire miscasts with a pure fireball, to remove the dependence
of Trog wrath on miscast code. A fireball scaled with player XL is a
fine "firey rage" to fit with the theme of Trog.