Adjust Cheibriados wrath
Replace miscasts in Cheibriados' wrath in two places. Miscasts are used
a as a source of "random bad stuff" and its better to have "thematic bad
stuff".

For "very high tension" retribution, add a lot of noise and then
fallthrough to the sleep and slow effects (to bring monsters to the
debilitated player).  For "low/no tension" retribution, deal some stat
damage.

The noise is given a clock flavour (Cheibriados strikes the hour.), and
the stats are in line with Cheibriados' gifts.