Rework Absolute Zero range and spell noise
Although this spell has some fundamental limitations compared to the
other level 9 direct damage spells given its single monster effect and
target randomization, it does feel kind of "cheap" at the highest
ranges. Deleting a monster that comes into view in a single turn
may be a valid effect for such a high cost spell, but doing this
repeatedly can be a bit dull.

This commit limits the range of Absolute Zero to 5 and restores the
original behaviour of having noise scale with spellpower. The spell
makes noise 40 effect at 0 spellpower, scaling down to 0 effect noise at
spellpower 200. The casting noise of 9 at the player's position will always
be present. Additionally we fix the spell noise display to properly
consider spellpower in the noise description.

This change doesn't prevent combinations of Vehumet, Darkness spell,
Robe of Night, and the Nightstalker mutation from achieving LOS range
for the spell, but getting those things in the right combination is at
least somewhat rare.