Make distortion tacticool
The teleportation effect was a "light instakill" for distortion, with a
variable impact from "a bit more than blink" to "that was a tough
monster I don't want ambushing me so now I've got to hunt it down".
Banishment works better as a true instakill, and the remaining weights
have been moved to the other distortion effects.

For comparison the old weights table:

33, SMALL_DMG
22, BIG_DMG
5, BANISH
15, BLINK
10, TELEPORT
10, TELEPORT_INSTANT
5, NONE

And the new weights table:

35, SMALL_DMG
25, BIG_DMG
10, BANISH
20, BLINK
10, NONE

This was not a proportional distribution of weight, and if the new
distortion becomes too strong with a 10% banish these numbers can be
further adjusted.