Work around a seed stability issue in regret_index_temple_of_mockery
This places different monsters depending on player species. The choices
it makes can affect rng differently depending mainly on the items that
come with some of these monsters (but potentially other factors as
well). This commit works around the problem by placing the mocking
monster when the player enters los, rather than immediately.

An alternative would be to use a subgenerator (perhaps for all vaults,
or for some with a specified tag), but there's not a great way to do
that that doesn't have a huge effect on many seeds that I came up wuth.
It still might be worth doing in trunk.