make merfolk pathfinding not mind about traveling through water
Formerly, merfolk pathfinding would count a tile of water as having
taking 2 turns to cross so that merfolk would avoid traveling through
water because 'their boots would meld'. In water merfolk gain bonus EV
and speed (equivalent to 4 boots of running), so this was a bit silly.

This commit makes merfolk pathfinding count a tile of water as 1 turn.