Remove autotravel water disincentive for merfolk
Formerly, merfolk pathfinding would count a tile of water as having
taking 2 turns to cross so that merfolk would avoid traveling through
water because 'their boots would meld'. This behavior interrupted
travel, but that travel interrupt was fixed in 771548fc26.

This commit makes merfolk pathfinding count a tile of water as 1 turn.

[ Committer's note: Closes #1343, reworded commit message to refer to
  the autotravel fix. ]