Adjust the scythe of Curses
Remove the dependence on necromancy miscasts, replacing them with death
curses. On wield the scythe delivers a severity 0 curse, and on hit a
severity min(damage done, 27) curse to monsters of natural holiness.

Also remove the item cursing. That was a flavourful annoyance more than
a serious limitation considering the quantity of ?rc available to
players.