A gauntlet map featuring an early escape option
The map gammafunk_gauntlet_77_escape_option features a linear path
through 5 7x7 arenas where the player may take a transporter to the
minotaur after each of the first four arenas instead of proceeding.
Arenas upgrade to tier 2 difficulty as early as the 3rd arena, with the
loot count likewise increasing. The average total loot from all arenas
is 8.75 items, which is a bit higher than the average of 7.5 for other
maps, but getting all items requires completing all 5 arenas. It's
fairly likely that the player will encounter a difficult arena and want
to exit to the minotaur instead, but the extra loot in the later arenas
might compel them to take a risk.