Make vault epilogues work again
This mostly affects messaging when entering wizlabs, and in cases where
a special panlord level generated as the first Pan level and this level
got generated out of order. (Could happen on full pregen, or if a player
entegered a zig before pan.)

Epilogues have been semi-broken since pregeneration. They are intended
to run after the player has entered the level, but before pregeneration
this always happened immediately after level generation with a guarantee
of full generation information kicking around. The bug is that if a
level is generated and not immediately entered, its vaults get saved to
disk and the epilogue is not serialized. On load, an incomplete map_def
is used for the vault information that is missing most things you'd need
for generation (rightly so), including the epilogue. So the epilogue
code tried to run, but was always empty in this case.

Rather than serialize the epilogue (which may be possible, but seems
like a bit of a minefield as I don't think we serialize any lua code
right now), the fix here is to run the epilogue from the current .des
cache's copy of the map. Obviously this could lead to save compat
issues, but this is exactly how lua code for markers is handled and the
issues are the same.