A Gauntlet map with a choke point
The map gammafunk_gauntlet_mixed_choke_point features mostly 7x7 arenas,
but with all 9 paths leading through a central 9x9 arena. The paths
start with a choice of 3 7x7 arenas that are tier 1 difficulty, leading
to the central arena that has a 50% chance to be tier 2, followed by a
choice of 3 7x7 arenas that are tier 2.