Fix icy wall descriptions
The icy walls made by Frozen Ramparts have a covering description for
the 'x' and 'x-v' commands that wasn't updated when the spell was
changed in c9638ee9a. Update the descriptions to refer to ice instead of
icicles, and expand the wall covering description to give the effect
details we provide for the spell.