Adjust wall selection for Frozen Ramparts (#1330)
This spell used a distance iterator to select walls in range, but that
iterator requires locations to be in bounds for player movement, meaning
at least two squares from the map border. This meant that the spell
couldn't affect otherwise normal walls adjacent to the map boundary, as
seen in #1330.

This commit switches to a LOS_NO_TRANS radius iterator, letting the
player affect these walls with Frozen Ramparts. The use of LOS_NO_TRANS
is a new restriction, but is appropriate since we don't generally allow
damaging monsters behind glass.

Additionally, this commit allows the spell to work on permarock walls,
since there's no reason why these should be excluded. Closes #1330.