Remove the Deflect Missiles spell
The RMsl spell was removed in 6993651381ec54b41ad219c4c6380028b2801db6,
and most of the arguments for its renewal also apply to DMsl. It's true
that it requires much more investment, but that also makes it much more
marginal: we carry around the code and design complexity of multiple
missile repulsion levels, and the notion of non-expiring "permabuff"
spells, for a spell that essentially duplicates an item property but
stronger, without it even being something that most characters will
have any opportunity or reason to learn.

After this commit, the design is crisper and easier to learn: there's
one level of missile repulsion and you get it from armour, gods, or
intrinsically; spells don't come into it.