Work around crash caused by early-game messaging
This works around a crash caused by a Gnoll skills check during game
resumption. If a gnoll had all skills at 27, a message was printed
before the game had been fully restored. The check changed in this
commit blocks layout changes until the game has been fully restored;
this also prevented the message window from being given a size. To work
around this, we explicitly give the message region an arbitrary size.

Unfortunately I can't remember which parts of the tilesdl code were
crashing or how to reproduce that issue. The underlying issue is that
the message handling code does its own screen output. But, as with all
things crawlcode, this is extremely non-trivial to fix.

Fixes #1346.