Refactor monster abjuration to be a true spell
Monster abjuration is automatically added to monsters with certain
summon spells, but the implementation was a bit of a hack. This refactors
the code to use a modern spell_logic struct for abjuration. It keeps the
check to abjure instead of using certain summon spells, in lieu of
adding SPELL_ABJURATION to a bunch of monster spell books, but the
codepath is now similar to legendary destruction instead of a different
form of special casing.