Clarify windows package build instructions
Currently INSTALL.md refers to the release guide for building packages
under MSYS2, which is not very clear and also doesn't mention the zip
and nsis dependencies. While users can figure this out, I think it would
be good to have it explained directly in the INSTALL file.

Additionally, the 'mingw64/' prefix in row 222 is not needed.