Tone down hell effect miscasts
A level 5, fail 50 miscast does 125 damage on average with the new
miscast formulas, and fail was being called with random2avg(97,3) for
hell effects.

This reduces hell effects to level 5, random2avg(40, 3) miscasts, which
do a maximum miscast damage of 100, and an average of 50. This will be
further reduced by a subsequent commit toning down miscasts over all.