Remove the Controlled Blink spell
It was already bumped to level 8 in 75a41fec5b260e3405678e92336f8d186af7b559
for balance reasons; while it's not much of a balance concern at that
level, it's still perhaps the most boring and predictable of the
high-level escape option spells, allowing easy escape from many
situations without the tactical or strategic concerns of spells like
Death's Door and Borgnjor's Revivification, as well as wholly
duplicating a consumable, unlike similar spells like Passage of Golubria
and Disjunction that have their own unique mechanics.

In addition, its relative spammability and complete lack of permanent
costs compared to the scroll also encourages the existence of arbitrary
and undiscoverable mechanics like the Zot/Orb translocation
restrictions. (Passage of Golubria is probably still too good on the orb
run to just remove those outright even after this change, but hopefully
future adjustments to that spell could make it possible.)

Passage of Golubria replaces it in the book of the Warp as the closest
equivalent spell.