Remove Passage of Golubria from the Warper book
It's a very strong and abusable spell to be giving away in a starting
book, and Warper is already one of the stronger backgrounds, especially
in terms of escape options (they start with a scroll of blinking!).

This also reduces the duplication between the starting book and the
book of the Warp, now that the latter also has this spell; they already
shared Gell's Gravitas. It also remains available in the book of
Burglary, as the spell fits well there and it's not a starting book.