Don't contaminate the player on controlled blink
It desn't seem necessary to have another limiting cost when it's only
available on a scroll, and the amount of contamination is basically
irrelevant (e.g. invisibility gives ~5000 contamination, unwielding
*Contam gives 7000). Contamination is also generally the domain of
spells, clouds, and traps rather than items, so it fits less well with
controlled blink now that it's a scroll-only effect.

If scrolls of blinking end up need nerfing (e.g. because of Orb status
removal), then a value like 1000 + random2(750) would be reasonable.