Refactor feature extra tiles handling
This commit groups data for feature descriptions together. Preivously,
titles and descriptions were separated from tiles.

This commit also fixes the display of the disjunction tile; it was never
a cloud, so the previous code never worked properly. Unfortunately, since
this code cannot access the packed_cell, we are limited to showing a
fixed tile for disjunctions, not the tile that is present at the
described map cell.