Add all_item_subtypes()
This commit replaces explicit iteration over an enumeration with the
use of a new helper function that encapsulates the intended behaviour,
making call sites more readable and less likely to break. The function
implementation is in a header file, which should allow the vector return
value to be optimized away.

This change permits refactorings removing old items from the game that
additionally add runtime checks for old items.