Guard against repeated calls to lua_object_gc (ruderubik)
This call is bound to the metamethod __gc, which can be called
arbitrarily from lua code (though it is always automatically called as a
finalizer before lua collects the memory allocated by lua_newuserdata).
This is a minimal change so that `delete` can't be called multiple times
on the same pointer. (All of this code is a bit of a mess.)

(cherry picked from commit 7f8d8a8652f7b62a65e1f4ebf3f194a4ccd17dac)