Pointer checks for lua userdata objects
This systematically adds checks to ensure that when userdata wraps a c++
object in some way via pointer, the pointer is not nullptr. On the dlua
side I'm just doing this for the sake of consistency and good practice,
but on the clua side this aims to prevent an attacker from using calls
to the `__gc` metamethod to get crashes on demand.

I kind of hate all these macros, it could probably use a refactoring to
make it less brittle.