Adjust monster behavior for working with ranged allies
Commit 6f887751 (in 2008 by jpeg) added behavior to have monsters of
animal intelligence or higher spill
out of corridors when attacking the player if followed by other
monsters; additionally it made intelligent monsters (at the time normal
or higher, now just human) check for an aligned monster with a ranged
attack further back and move out of the way. The example from the commit
message:

    ####.
    .rgo@
    ####.

    will within two turns change to

    ####g
    ...r@
    ####o

In that logic this movement only triggered if walls were in the way,
even though opening up lines of fire for ranged attackers is a good
thing to do in the absence of a corridor.

This commit changes the behavior to only check the adjacent following
monnster in the corridor situation, and check for ranged allies
regardless of the presence of walls in the two spaces marked with ?

    ###?.
    c..o@
    ###?.

    has a chance of changing to

    ###?.
    c...@
    ###?o

if the orc moves before the centaur, regardless of the features at the
?. The decision to spill out was given a coinflip later in the
development history (hard to git blame because it moved from monstuff.cc
to mon-act.cc), and the choice of up or down is also randomized.

In the monster intelligence simplification of d63ab3c5 (in 2015 by
PleasingFungus) the intelligence component of the ranged ally check was
removed, but not mentioned in the commit message as an intended effect.
This is restored, so that animal intelligence monsters can be used for
"cover" of sorts.

Finally, this commit has monsters treat the player as a "ranged
attacking ally" for the purposes of this check. It's not a sophisticated
"stay out of my line of fire" behavior requested by some players and
might negatively effect players who use their Hepliaklquana ancestor to
block while polearming. A more sophisticated behavior could be added,
but without adding an ally command (ugh) or some complicated guessing
about player behavior (also ugh) it's not clear what the right move is.