Change explore_auto_rest default to true
For new players, it's a much better experience for this option to be
enabled, to prevent accidentally auto-exploring at low HP into an orc
priest or similar.

Players aiming to maximise score will want to disable this, and
potentially also Ghoul/Troll characters (although it's not a major
issue past the first couple of floors). But on the whole, enabling by
default seems to improve more games.