Make allies annoyed by translocations other than Recall
Lesser beckoning working on allies offers opportunities for tedious ally
micromanagement of a kind we'd rather not allow; the targeter already
doesn't think the spell applies to allies, but if it was used on them it
would work without angering them. This change is partly to remove this
micromanagement, and partly for consistency with the change to teleport
other.

Teleport other on allies could be used as a tedious combination with
zombies or summons and inner flame to seed a level with bombs from the
safety of the stairs. Awake hostile monsters would engage the summon or
ally and suffer explosion damage, possibly resulting in a kill.