Permit inner flame on summons
Commit 3fe85ef3 allowed immolation scrolls to affect summons; the
previous commit removed a degenerate interaction with teleport other.

So that the two effects are consistent, this allows casting inner flame
on summons. If "ready" access to suicide bombers becomes an issue, inner
flame can be made to annoy allies instead. However, this was already
possible with skeletons and sticks to snakes and outside of "snakebombs"
created with immolation scrolls players are not going around casting
inner flame on allies in place of using fireball.