Fix map cell corruption on unmarshalling (#12218)
When marshalling flags, the unsigned flags are implicitly cast to a
signed integer of the size being marshalled, but during unmarshalling no
cast was performed, leading to values with the high bit set undergoing
sign extension. This resulted in visual artefacts.

I've added casts to the marshalling code, not because they are
necessary, but because I don't like the implicit behaviour and want to
make it clear to the reader that this is deliberate.