Enable deep elf elementalist tile animation
When the new tiles from roctavian were added in 16654a8,
the elementalist cycle animation wasn't added, due to wielded weapon
not displaying correctly.

Aidanh found the reason, it's that all the variant tiles need to be
listed in get_weapon_offset in tilemcache.cc.

This commit enables the animation, and also reuploads the tiles,
because only 3 out of the 4 originally submitted tiles were uploaded
in the previous commit, and they need to be in the correct sequence.

I tested it in a local windows tiles build only, working fine here.