Fix high CPU usage on Windows builds (Goratrix, advil)
The timer callback should return an interval for the next call, but was
erroneously returning 1 (from the SDL_PushEvent call), where instead
it should have just been returning zero, as the callback did prior
to a840853. The return value of 1 caused the callback to be called
repeatedly every 1 millisecond, causing crashes and high CPU usage.