Allow monsters to place on traps (#1283)
Traps are triggered by moving into them. Forbidding monster placement on
traps dates back to a simulationist restriction on monsters placing on
teleport traps. The restriction that traps only fire in sight of the
player means we've given up on that illusion.

The code comment indicates the restriction was to be limited to teleport
traps, but in actuality applied to all traps, leading to frustrations
with animate dead in spider and unexpected limitations on summoning.

Restricting to a special case just for translocational traps seemed
unnecessarily complicated, so this commit permits placement on all
traps.

[ Closes #1283 ]