Prevent inadvertent opening of ebering_ghost_funeral_grove (#1077)
In Swamp this vault blends in and a player might fireball it open not
realizing it's a ghost vault. Monster tornado is highly unlikely to
interact with this vault (but might as a rare Zot trap effect or if
somehow an air elemental comes in), but again preventing inadvertent
opening is important for the depth of this ghost vault.

[ Closes #1077 ]