Change show_more default to false
For new players, it's confusing and inconsistent to sometimes be
prompted to press space while waiting to play your next turn.
Additionally, the prompt to press space is small and easily overlooked
(and it's not clear what the --MORE-- prompt means unless you are
familiar with Unix command line environments).