Remove targeting from Summon Lightning Spire
SLS creates a tough, immobile ally that the player can manouver around
to gain maximum benefit from. However, the initial placement was
targeted, which provided a way to gain maximum benefit with no cost
other than the invested MP.

Removing the targeted construction is a nerf. It is also in-line with
the guiding ideas behind positional magic: that tactical benefits from
spells should be gained by movement rather than targeting. The strength
of an optimally positioned spire remains the same, but the cost of
obtaining that position now includes some amount of movement instead of
only MP and targeting time.