Prevent Eldritch Tentacles from attacking out of LOS
The summon-los-restriction dates to a9662255 in 2012. The commit message
seems like this is intended to apply to all summons; however, eldritch
tentacles are created with an abjuration duration of 0, so the code
introduced at that time did not apply the restriction to them. Four
years later, in 2016, the restriction was added to FAKE_ABJRATION
allies. This too did not contrain the tentacle, which is timed out with
ENCH_PORTAL_PACIFIED.

In the 8 years since a9662255 this oversight or bug has become an
accepted quirk of the spell and not treated like a bug. It is also a
source of out-of-los attacks of the kind we intentionally prevent. So,
in a long overdue commit this finally constrains malign gateways to not
attacking out of LOS.