Reduce the delay before an eldritch tentacle enters
The previous delay was set after the old out-of-los attacking
behavior had been accepted and chosen with the old behavior in mind. Now
that the player must stick around for the tentable to work, bringing it
out in 1-2 turns instead of 2-4 is appropriate. This also increases the
threat of Zot-trap opened gateways.