Fix description of what happens when attacking summons
Since 3db639d1 it has been, from the user's perspective, impossible to
kill most friendly summons -- they are simply dismissed on an any damage
or other annoyance.  This is particularly important for something like
lightning spire where dismissing the summon may be tactically useful.
So, improve the description in an attempt to make it clearer what will
happen if you anger such a summon.