Script to assess removedness
I became interested in the question of how much, quantitatively, gets
removed and added from each vesrion. This script is able to provide
counts for various kinds of things (species, branches, spells, item
types) in crawl and will work on versions going back to 0.3 (with a few
caveats before around 0.10).