Declare some tags functions in tags.h
These functions are now integrated into unit tests, hence must be
non-static. Before this commit, they lacked declarations in tags.cc
itself and had declarations in a unit test file, but this generated
compilation warnings for tags.cc. Per discussion with aidanh, it's
probably a good thing if the unit tests are forcing things that need to
be tested to have header declarations, since it helps clarify program
structure. So we move the required functions' declarations to tags.h and
remove the duplicate declaration from the unit test files.