Standardize conditions for including unistd.h
The immediate target of this is a report of compilation failure on
freebsd, but I noticed that we are all over the place in how this is
conditioned. Because several of these were non-windows defines, it's
possible that this could break compilation on obscure/non-supported
build targets like hurd(!). Probably this should be some kind of actual
posix check...

(cherry picked from commit 1b96fe49b65ee3cc1e82f53194882713db3ac5f1)