Refactor mons_condition_string
The previous commit introduced some new architecture, but had extra
logic woven in that isn't needed (the newcondition/listall boolean
parameter was only false when writing a description of the scene in the
chardump, but it's fine to be verbose there). This commit simplifies the
"vector of structs with boolean test functions" to a map from
monster_info_flags to their singular and plural display names. A
subsequent commit will refactor monster_info::atributes() to use the new
structure to unify the two description lines.