Give a chance for more than one abyss xp exit to spawn
The fact that the first one is an exit, and everything after that is
guaranteed to be a downstair, is very confusing to players / spoilery.
(At an extreme, I've seen multiple reports over time of players thinking
each time the stairs generated that they simply got unlucky, and
continuing to fight to generate more, sometimes over and over again.
Worst for qazlal players.) This commit changes things so that after the
initial exit, there's a 50% chance for downstairs vs exit. The xp ramp
continues until you exit the abyss. I considered other possibilities
that were less of a change like resetting the bool on level change, but
they were somewhat spoilery too.

I don't think there's a balancing issue here, but if anyone else does,
to get closer to the previous behavior you could adjust the
probabilities as the XP threshold for spawning increases.