Don't let abyss xp stairs overwrite exits
For some reason, DNGN_EXIT_ABYSS is not a stair, and not even a branch
exit technically. I suspect another past source of UI confusion with xp
exits was that a player could overwrite an xp-generated exit with down
stairs by fighting more, and then never get another exit. (Whereas an
exit could not overwrite a stairway, though that wouldn't have come up.)
This also prevents xp exits from overwriting a few other things, such as
transporters. It could be worth giving a message past the xp threshold
in order to cue the player that they need to move if they want something
to generate?