Add optional player-relative coordinates to the Monster/Item view
This was requested as an accessibility feature for players using screen
readers, as the ^x screen is much easier for them to use.

This treats the up direction as positive y, even though crawl's grid
uses screen coordinates. However, a standard right-handed cartesian
coordinate system is more familiar to most users who are not graphics
programmers, so the minus sign is worth it.