Decouple key input from targeting_behaviour
targeting_behaviour has absolutely no business getting user input
directly; it should only encapsulate behaviour for a given targeter. How
a given targeter responds to a given key event (i.e. the mapping of a
key to a command) is fair game, but how that key event is received is
not. This was never used in practice, thankfully.

This will allow refactoring the direction chooser view with widgets,
and also permit testing individual targeting behaviours, since they're
no longer so tightly coupled to the input system.