Consolidate directn shadow-step check
I don't entirely understand this code, but the way the 'main loop'
is set up, the target is always changed through set_target() before
any key input will be processed; so there should be no change to
valid_shadow_step between when it is set and when it is checked. This
means that this shadow step code can be moved together.

This makes the set_target() method very simple (and ready for removal),
and it gets rid of a bit of state on direction_chooser.