Unstack tremorstones and make them indestructible
Is this another circle on the merry-go-round of "damaging purple
evocable that also does self damage" a la tome of destruction and disc
of storms? Time will tell. Tremorstones' effect is much simpler than the
two mentioned items: randomized but possible to reason about without
consulting a spoiler for the chart of switch statement cases.

Now that tremorstones don't have a targeter and scale with evocations,
its important to give the player an opportunity to get a feel for the
self damage. The self-damage also tactically limits the effect without
contributing to power spiral in the same way as a consumable. Because
the effect scales with evocations, there is a strategic cost in getting
greater benefit from tremorstones. For all of these reasons, making them
consumable no longer fits.

If the self-damage is not enough of a brake against spamming, XP-based
charges can be added. (Indeed, moving all evocables to XP-charging would
be a good idea to help with power-spiral, but it's too close to 0.25 to
do that now.)