Fully sync player on webtiles load (12117)
To determine what values to sync, all the `_update_blah` calls require
that the initial values it is provided with are already in sync with the
webtiles server (they only sync changes unless forced). However, the
player_info class does not initialize most of its values, and the first
instance of _send_player would typically be triggered on a redraw, which
does not force sync. So when the on-load player_info value of some
setting matched its initial value (which is undefined / up to the mercy
of the compiler), that value would not get synced until it changed, or a
spectator connected. This commit ensures that the first instance of
_send_player is always a force sync, i.e. sends every value. There are
other ways of fixing this but this one seemed like it would be
reasonably safe.

This was manifesting in a very specific way in 12117, where a quiver
item on the `a` slot translated to a value 0, which in several build
settings turns out to be what an uninitialized uint8 gets set to --
leading to the quiver value not getting synced in this one specific
case. However, I wouldn't be surprised if there were other small
effects.

This issue might conceivably arise for other initial sync things; it
makes me wonder why we call _send_everything for spectators but not ever
for the player?