Fix an iterator validity issue in sprint_mu (12192)
When the door property called its veto check and removed the marker,
that invalidated the iterator used in map_markers::property_at, leading
to a crash. I think this code has been broken for ~ 6 years? (The
previous version used a const iterator that I think would be resilient
to this.)