Fix a theoretical terrain change crash
I can't replicate this, but the previous implementation should be able
to crash in the following way: if there are two terrain changes of the
same type on a single position (something that is possible in
principle), then reverting the terrain changes at that position would
delete all markers associated with that position at once. Then, the
iterator over markers will have a deleted marker in its future. Instead,
save up all positions with markers to be reverted, and do them last by
position, outside of the marker iteration.

I *think* this might be able to happen in complicated circumstances
involving water nymphs over stairways and lugonu corrupt. But, that is
just a theory.