Add preservation (rCorr) and stealth to cloaks
The previous commit made the cloak ego list a bit thin. This commit adds
preservation and stealth as possible cloak egos.

"preservation" is reintroduced as an armour ego ego, but now it is just
a name for rCorr (the player-facing name could change, but "cloak of
resist corrosion" just doesn't have the same ring).