Remove scarves of spirit shield
Offering guardian spirit on a different slot in non-unrandart form
offered such power combos as spirit+faith and spirit+acrobat at the
price of making the scarf ego list overlap with the amulet ego list.
With invisibility moved to scarf egos, there are enough scarf egos that
this sort of padding isn't needed.