Add Harm to scarves
This adds Harm as an armour ego, without the drain penalty on removal.
The Drain penalty was to prevent swapping in urgent situations, but in
practice players just ate the skill drain to get out of Harm when
escaping. Putting it on an armour slot makes using Harm a much stronger
commitment.