Remove amulets of reflection
A small plus to SH and the reflection property was a minor boost in most
cases and easily outshone. Newer players enjoy the satisfaction
of the reflection ego but to a point where it becomes a newbie trap:
they'll take reflection over much stronger (though more subtle) amulets.

They also suffer from a tedious-optimality: it's technically always good
to carry one when possible while using an amulet of guardian spirit, and
swap to it should your MP pool empty; or when using an amulet of rage.
The previous amulet reform aimed to make the amulet slot one that the
player commits to well in advance of a given fight, but the presence of
a slew of older amulets hampered that goal.