Remove amulets of Rage
The remaining evocable active amulet, they no longer fit with the
remainder of the amulet slot. Lacking a swap penalty, an amulet of rage
can be swapped to by a character using a different amulet's effect when
it's not needed if berserk can finish an encounter, since most current
amulets swap penalties come on activation, not removal.

Berserk is a powerful effect, unlimited berserk more powerful still, so
I don't see a desparate need to move +Rage to a scarf. In doing so,
+Berserk artefacts and Trog become a bit more special.