Don't partly override cell cache with global data
In d8832a91e, global state was more fully cached in the packed cells
later used for Tiles rendering. This included the tile flavor data used
to describe floor tiles of features. However the WebTiles code that
handles tile flavor was not updated, and it was subsequently overriding
the packed cell cache.