Refactor skill experience init/finish handling
SkillMenu::init() is regrettably called for each reallocation of the
skills menu wrapper widget. This caused init_experience() to be called
multiple times, but only the first call had any effect. So I've hoisted
this call to outside the SkillMenu.

exit() always called clear(), but this was called by the wrapper
function running this menu anyway.