Add experimental save info to webtiles lobby
This adds some features for showing save info to the webtiles lobby, a
longstanding ui gap. This is all relatively experimental and disabled by
default; I think it may need some tweaking once it hits a live server.
Because this feature is aiming for backwards compatibility there are a
number of idiosyncracies to the implementation. Number one is that it
accepts the (extremely confusing) way dcss treats save slots - for each
player and game version, one save slot for sprint, and one for all other
game types.