Fix a bug with shared save slots in the lobby
When starting a new game or ending a game, save slot info did not get
updated properly for other game types that shared a slot. This is a sort
of basic fix, in that it isn't caching very well (because the way crawl
handles save slots is a bit of a mess).

No servers are using this yet, so I don't think it has been an issue in
practice.