Revise transmutation miscasts
The polymorph effect was very flavorful, but even at minimal power a 150
aut badform was the result. Compared to other miscasts this was too
intense. Reducing form duration would require a significant overhaul of
player badform durations and power scaling because of how the existing
codepath is set-up.

Instead, since this is close to 0.25, this commit changes transmutations
miscasts to increase player contamination: double existing contamination
and then add the resulting miscast damage to the contamination score.

Contam is still flavorful (since transmutations change the body and
contamination is residual magic in the body and high contamination can
cause mutation), though not as dramatic as polymorph.

Possible idea for 0.26: add polymorph as a possible consequence of
decontamination.