Remove blood scent
Blood scent was a very obscure mechanic. As far as I can tell, it is not
revealed explicitly in game at all. Since it primarily affects how
certain out-of-los monsters might notice and track you, it is very hard
for an unspoiled player to notice the mechanic by playing. It is
attached to a feature, blood spatter, which is easily assumed to be
purely cosmetic. It is flavourful, but that's not much good if no one
knows about it; and even if it were revealed (say, in monster
descriptions), it is so minor that likely most players would ignore it
anyway. Mechanically, the gameplay it produces overlaps considerably
with noise.

This commit removes the blood scent flag and the logic handling the
property.