Simplify some pathfinding code
Commit d63ab3c5 ("Simplify monster intelligence") made it impossible for
mons_tracking_range to return a value of 0 (signifying no range
restriction, for highly intelligent monsters). Therefore there is no
need for conditionals checking to see if that value is greater than 0,
and this commit removes some of the special handling for that
case. (Probably a more thorough pass can find more of this.)