Remove web sense
Many of the remarks in the previous commit removing blood sense apply
here. Web sense is in a somewhat better place (it's attached to a
clearly mechanically relevant feature, and it's at least revealed in the
in-game description for webs). But its gameplay effect is mostly to
incentivize something that doesn't need further incentive, namely, not
stepping into webs when monsters might be nearby. And it is still a
spoilery mechanic with a minor gameplay effect that dubiously justfies
its added complexity.

This commit removes the web sense flag and the logic handling the
property.