Temporary check to bridge a change in client.js
client.js is served in a way where its browser cache isn't invalidated
by changes, and so the lack of an updated version was causing many
problems with this line.  Kind of ugly, but having this around for a
while is simpler than dealing with a lot of people having trouble with
browser caches.

(client.js should probably have some kind of versioning so that changing
it doesn't go so poorly?)