Add a global scaling option for local tiles
OS support for high-dpi rendering varies widely, but the sizes of UI
elements are very hard to adjust for higher-resolution displays where
high-dpi is not supported. This commit adds a coarse global scaling
parameter that mimics high-dpi behavior, scaling by a value from 1-8.
I'm not sure if there's any current devices out there that would need 8,
but 2 and 3 work reasonably well for my 2k ultrawide monitor, so I
expect that even 4 will be usable for a 4k monitor without high-dpi
support.

This is mostly implemented at a very low level by exploiting existing
high-dpi code; it would be nice if this could be consolidated to be done
entirely at the `WindowManager` level.