Restore map view clamping behavior (elliptic)
ca36f614e438 conditioned clamping on `map_bounds` instead of
`known_map_bounds`, the behavior before this commit. I think this was
unintentional, and it leads to weird behavior on unexplored levels in
console (e.g. lets you move the cursor outside of the screen, restores
to known_map_bounds when you move outside of regular map_bounds). This
commit restores the previous behavior for `X` where you can't scroll
beyond known map bounds. As a side note, this is inconsistent with `x`,
where you can, which is kind of messy.

TBD how this commit interacts with the test in bd17af9.