Document python-is-python3 for Ubuntu
We are stuck in a limbo of supporting ancient webtiles servers with
Python 2 installed only as 'python', and also newer OSes where Python 3
is only available as 'python3'.

Ideally the Makefile can try all three names for Python automatically
and use the first it finds. But until someone adds this behaviour we
should document the correct approach.