Tweak how seed games are displayed in ?v and chardump
Consolidated code between these two. This doesn't matter so much for
chardump, where the first line also says "Seeded", but it's a good
reminder for players looking at their seed in ?v. I've heard of players
forgetting they were starting a game in seeded mode (e.g. when setting
it via the rc) so there's maybe more telegraphing of this mode that
could be done somewhere.