Show seed for seeded play with welcome message
This is because of reports of players forgetting that they had a seed
set in their rc file. This also prints it on the "welcome back" message,
not sure if that's too repetitive.

Also, tweak this message a bit.

(cherry picked from commit 00fb324cd09f9518cf33ae1896d3cc5e93489a07)